# chore: 3x speedup, 0.33x memory usage, one line :) (SE model)